refactor(WebAPI): Fix many warnings

+++ tests/UnitTests/Application.UnitTests/PlexDownloads/Execute/DirectDownloadClient/EnsureDownloadDirectoryCommand.UnitTests.cs
@@ -106,8 +106,8 @@ public class EnsureDownloadDirectoryCommandUnitTests : BaseCommandUnitTest<Ensur
     {
         // Arrange — configure the drive with less free space than the requested file size
         const string directory = "/downloads/reaparr/Movies/Test Movie (2024)";
-        const long fileSize = (long)ByteSize.BytesInMegaByte * 500; // 500 MB
-        const long availableSpace = (long)ByteSize.BytesInMegaByte * 100; // 100 MB — not enough
+        const long fileSize = ByteSize.BytesInMegaByte * 500; // 500 MB
+        const long availableSpace = ByteSize.BytesInMegaByte * 100; // 100 MB — not enough
 
         var command = new EnsureDownloadDirectoryCommand(directory, fileSize);
 
@@ -137,7 +137,7 @@ public class EnsureDownloadDirectoryCommandUnitTests : BaseCommandUnitTest<Ensur
     {
         // Arrange
         const string directory = "/downloads/reaparr/Movies/Test Movie (2024)";
-        const long fileSize = (long)ByteSize.BytesInMegaByte * 500; // 500 MB
+        const long fileSize = ByteSize.BytesInMegaByte * 500; // 500 MB
 
         var command = new EnsureDownloadDirectoryCommand(directory, fileSize);
 
@@ -156,7 +156,7 @@ public class EnsureDownloadDirectoryCommandUnitTests : BaseCommandUnitTest<Ensur
     {
         // Arrange
         const string directory = "/downloads/reaparr/Movies/Test Movie (2024)";
-        const long fileSize = (long)ByteSize.BytesInMegaByte * 500; // 500 MB
+        const long fileSize = ByteSize.BytesInMegaByte * 500; // 500 MB
 
         var command = new EnsureDownloadDirectoryCommand(directory, fileSize);
